✨ Init stateMachine for menu & game
