+++ include/clientRun/ClientRun.hpp
@@ -43,3 +43,5 @@ void update(Game &game, Window &window);
 void handleCommand(std::string command);
 void updateImagePlayer(Player &player);
 void handleMaxMin(sf::Vector2f &position);
+void updateGame(Window &window, Game &game);
+void updateMenu(Window &window, Game &game);
