+++ src/client/client/DataManager.hpp
@@ -6,6 +6,11 @@
 #include "client/graphic/Player.hpp"
 
 class DataManager {
+   public:
+      enum GameState {
+         MENU,
+         GAME
+      };
  private:
     int gravity;
     int speed_x;
@@ -18,8 +23,10 @@ class DataManager {
     std::vector<std::unique_ptr<Player>> players;
     sf::Texture texturePlayer;
     sf::Texture textureBackground;
+    GameState state = MENU;
 
  public:
+    std::mutex mutexState;
     static DataManager *instance;
 
     DataManager();
@@ -31,7 +38,7 @@ class DataManager {
     int getPort() const;
     std::string getIp() const;
     std::vector<std::unique_ptr<Player>> &getPlayers();
-    Player &addNewPlayer();
+    GameState getState() const;
 
     void setGravity(int gravity);
     void setSpeedX(int speed);
@@ -40,7 +47,9 @@ class DataManager {
     void setDebug(bool isDebug);
     void setPort(int port);
     void setIp(std::string ip);
+    void setState(GameState state);
 
     sf::Texture &getTexturePlayer();
     sf::Texture &getTextureBackground();
+    Player &addNewPlayer();
 };
